requirements.txt
pydantic==2.1.1
twscrape==0.17.0
python-decouple==3.8